pyproject.toml
[build-system]
requires = [
  "hatch-vcs>=0.3",
  "hatchling>=1.12.2",
]
build-backend = "hatchling.build"

[project]
name = "apophenia"
authors = [
    {name = "Hervé Beraud", email = "herveberaud.pro@gmail.com"},
]
description = "Extract and structure all the data from a Git repository to make them usable in RAG."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
url.Homepage = "https://github.com/openuado/niet"
url.Source = "https://github.com/openuado/niet"
url.Tracker = "https://github.com/openuado/niet/issues"
classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Development Status :: 3 - Alpha",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development",
    "Topic :: Utilities"
]
dynamic = ["version"]
dependencies = [
    "sentence-transformers",
    "gitpython",
    "faiss-cpu",
    "numpy",
]

[project.optional-dependencies]
dev = ["black", "isort", "pip-tools", "build", "twine", "pre-commit", "commitizen"]

[project.scripts]
apophenia = "apophenia.main:main"

[options.packages.find]
where = "apophenia"
exclude = "tests*"

[tool.hatch]
version.source = "vcs"

[tool.isort]
profile = "black"
